requirements.txt
Django==5.1
pandas
scikit-learn
joblib
keras
biopython
nglview
igfold